extend process.sh to support patch stages
The idea is that we want to support incremental tests where we do some
base build step and then apply a series of patches. We want to track the
compilation time after each of these patches separately. The model is
that from one directory $DIR we will create N files, one for each
patch. They will be named $DIR@$PATCH -- they are basically independent
tests from the POV of the front-end.

This is achieved by extending the makefile protocol mildly. There is now
a new target `make patches` that emits a list of patch names. These
should include the separator (e.g., `@foo`, `@bar`).

If the list of patches returned is empty, the code works as before.
We execute '(make all; make touch)+ make clean'. The results from
'make all' are stored in `$dir--$DATE--$i`.

If the last of patches is *non-empty* (say '@A @B'), then we execute
'(make all@A; make all@B; make touch)+ make clean'. The results from
'make all' are stored in `$dir@A--$DATE--$i` and `$dir@B--$DATE--$i`.

The naming convention for variants is as follows:

    @NNN-text

where NN is 000, 010, 020, 030, etc, and text is some useful description.
The @ is a distinctive separator we can special-case in the front-end
later.