modify compare.py to consider patches
This is not smart enough to track the time for each incremental change
separately. I didn't do that because I think it'd be better to retool so
that both `process.sh` and `compare.py` are using the same code (i.e.,
so we are measuring the same thing that the server is).